refactor fn to make easier testing
